Add XML with kernel EOL information to enable enforcement
kernel-lifetimes.xml is a mirror of the source of truth for kernel End-of-Life (EOL) for GKI development and release kernels. The fully up-to-date source of truth will be maintained at main:kernel/configs/kernel-lifetimes.xml

This version is for STS testing purposes and will be updated every six month and as needed.

Bug: 338636324
Change-Id: Idce8d88abbecc96f2010ec0911b141b40bbfead1
Merged-In: Idce8d88abbecc96f2010ec0911b141b40bbfead1
Signed-off-by: Todd Kjos <tkjos@google.com>